requirements.txt
argparse~=1.2.1
setuptools~=44.1.1
configparser~=4.0.2
Faker~=3.0.1
kafka-python
pytest~=4.6.11